Created util.go and moved some of the generic stuff there
